+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename `katja_metrics` to `katja_writer` and `katja_queries` to `katja_reader`
 * `katja:send_event/3`, `katja:send_events/3`, `katja:send_state/3`, `katja:send_states/3` and `katja:send_entities/3` can be used to force a transport (supported: `auto`, `udp`, `tcp`)
 * A default message transport (affecting only entities that are sent to Riemann) can be set using the `transport` configuration option
+* Add `katja:start/0` to start the Katja application and all of its dependencies
 
 ## 0.4